+++ R/flob.R
@@ -1,15 +1,16 @@
 #' Flob File
 #'
 #' Converts a file into a flob.
+#'  Flobs are useful for saving files in databases.
 #'
 #'  A flob is a file that was read into binary in integer-mode as little endian,
 #'  saved as the single element of a named list
 #'  (where the name is the extension of the original file)
 #'  and then serialized before being coerced into a blob.
 #'
-#'  Flobs are useful for saving files in databases.
-#'
 #' @param path A string of the path to the file.
+#' @param name A string of the name (without the extension) for the flob.
+#' If "" (the default) then the original file name is used.
 #' @return A named flob of the file.
 #' @seealso \code{\link{flobr}}
 #' @examples
@@ -17,8 +18,9 @@
 #' flob <- flob(path)
 #' flob
 #' @export
-flob <- function(path) {
+flob <- function(path, name = "") {
   check_string(path)
+  check_string(name)
 
   if (!file.exists(path)) stop("file '", path, "' does not exist", call. = FALSE)
 
@@ -26,11 +28,13 @@ flob <- function(path) {
   flob <- readBin(path, what = "integer", n = n, endian = "little")
   flob <- list(flob)
   class(flob) <- "exint"
-  names(flob) <- tools::file_ext(path)
+
+  names(flob) <- name(name, path)
 
   flob <- serialize(flob, NULL)
   flob <- list(flob)
-  flob <- as.blob(flob)
+  flob <- as_blob(flob)
+  attr(flob, "ptype") <- NULL
   class(flob) <- c("flob", "blob")
 
   names(flob) <- path
@@ -42,30 +46,40 @@ flob <- function(path) {
 #'
 #' Converts a \code{\link{flob}} back to its original file format.
 #'
-#' If path does not include an extension then it is added automatically.
-#' If path does include an extension and then it must match the flob's extension.
+#' If path ends with a file separator then the name and extension of the flob is added.
+#' If the flob is an older flob that was not saved with a name then it is named 'file'.
+#'
+#' If path ends with a file name without an extension then the extension of the flob
+#' is automatically added.
+#' Alternatively if path also includes an extension
+#' then it must match the flob's extension.
 #'
 #' @param flob The \code{\link{flob}} to save to file.
-#' @param path A string of the path to the new file.
+#' @param dir A string of the path to the directory to save the file in.
+#' @param name A string of the name (without the extension) for the file.
+#' If "" (the default) then the original file name is used.
+#' @param ext A string of the extension for the file.
+#' If "" (the default) then the original extension is used.
 #' @return An invisible string of the path to the saved file.
 #' @export
 #' @examples
-#' unflob(flob_obj, tempfile())
-unflob <- function(flob, path) {
+#' unflob(flob_obj, tempdir())
+unflob <- function(flob, dir = ".", name = "", ext = "") {
   check_flob(flob, old = TRUE)
-  check_string(path)
+  check_string(dir)
+  check_string(name)
+  check_string(ext)
 
   flob <- unlist(flob)
   flob <- unserialize(flob)
-  flob_ext <- names(flob)
-  path_ext <- tools::file_ext(path)
-
-  if(identical(path_ext, "")) {
-    path <- paste0(path, ".", flob_ext)
-  } else if(!identical(path_ext, flob_ext))
-    stop("path extension must match '", flob_ext, "'", call. = FALSE)
+  names <- names(flob)
 
   flob <- unlist(flob)
+
+  if(identical(name, "")) name <- file(names)
+  if(identical(ext, "")) ext <- ext(names)
+
+  path <- file.path(dir, paste(name, ext, sep = "."))
 
   writeBin(flob, con = path, endian = "little")