+++ NEWS.md
@@ -1,3 +1,14 @@
+# flobr 0.2.0
+
+By default flobs now include the original file name with the extension. 
+
+To read with a different file name use the name argument to `flob()`.
+The saved file name can be accessed using `flob_name()`.
+
+To write with a different file name use the name argument to `unflob()`.
+
+If the flob was read using a previous version of flobr which means it lacks a file name then by default it is written with the file name "file".
+
 # flobr 0.1.0
 
 * Added `flob_obj`